Remove obsolate files
This commit removes files merged in commit:
3447e138e9638f4afdfdbca522d5a010b3c85cf2
