+++ src/main/Main.java
@@ -7,8 +7,6 @@ public class Main
 {
     public static void main(String[] args)
     {
-        Database d = new Database();
-        UserPanel userPanel=new UserPanel(1);
         Application.appMain();
     }
 }
