+++ src/main/Panels/OfferPanel.java
@@ -1,0 +1,4 @@
+package main.Panels;
+
+public class OfferPanel {
+}
