+++ src/main/Panels/UserPanel.java
@@ -1,5 +1,6 @@
 package main.Panels;
 
+import main.Application;
 import main.Database.Database;
 
 import javax.swing.*;
@@ -11,24 +12,23 @@ import javax.swing.border.TitledBorder;
 
 public class UserPanel  extends JPanel {
 
-    public UserPanel(int Id) {
-
-        id = Id;
+    public UserPanel()
+    {
         init();
     }
+
     public void init()
     {
         int i;
-        final var iconPath = Path.of(assetsDir.toString(), "user.jpg");
+        final var iconPath = Path.of(Application.assetsDir.toString(), "user.jpg");
         ImageIcon image = new ImageIcon(iconPath.toString());
         JLabel label = new JLabel(image);
         GridBagConstraints c = new GridBagConstraints();
         Vector<String> header = new Vector<>();
-
         label.setBorder(BorderFactory.createBevelBorder(1));
         setLayout(new GridBagLayout());
 
-        data = db.getUserData(id);
+        data = db.getUserData(Application.getInstance().getUser());
         String[] titles = {"imie:","nazwisko:","email:","numer telefonu:"};
 
         ArrayList<JLabel> dataLabels = new ArrayList<>();
@@ -37,8 +37,7 @@ public class UserPanel  extends JPanel {
             dataLabels.add(new JLabel(titles[i]+data.get(i)));
 
         }
-        offers = db.getUserOffers(id);
-
+        offers = db.getUserOffers(Application.getInstance().getUser());
         c.fill = GridBagConstraints.HORIZONTAL;
         c.gridx = 1;
         c.gridy = 0;
@@ -75,12 +74,11 @@ public class UserPanel  extends JPanel {
                 TitledBorder.CENTER,
                 TitledBorder.TOP));
         setVisible(true);
+        setBounds(0, 0, 1000, 630);
     }
 
     private ArrayList<String> data;
     private Vector<Vector<String>> offers;
-    private int id;
-    public static final Path assetsDir = Path.of("assets");
     static Database db = new Database();
 
 }
